+++ contrib/datadog/tracing/go.mod
@@ -1,4 +1,4 @@
-module go.temporal.io/sdk/contrib/datadog
+module go.temporal.io/sdk/contrib/datadog/tracing
 
 go 1.16
 
@@ -7,4 +7,4 @@ require (
 	gopkg.in/DataDog/dd-trace-go.v1 v1.42.0
 )
 
-replace go.temporal.io/sdk => ../../
+replace go.temporal.io/sdk => ../../../
