+++ contrib/datadog/tracing/interceptor_test.go
@@ -1,4 +1,4 @@
-package datadog
+package tracing
 
 import (
 	"testing"
